requirements.txt
# Wiki base model tooling
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytest>=7.0.0
PyYAML>=6.0
